chore: Put new endpoint in own resource
